fix(pay): adjust payment utils
