+++ app/confirm-order/choose-receipt/page.tsx
@@ -28,11 +28,7 @@ const ChooseReceiptPage = () => {
       <Title title="選擇收件人資訊" goBackUrl="/confirm-order" />
       <div className="flex min-h-screen w-full flex-col items-center bg-default">
         {addresses
-          .filter((opt) =>
-            deliveryType === 'home-delivery'
-              ? opt.LogisticsSubType === 'home-delivery'
-              : opt.LogisticsSubType !== 'home-delivery',
-          )
+          .filter((opt) => opt.LogisticsSubType === deliveryType)
           .map((opt) => (
             <ReceiptInfo
               key={opt.id}
