+++ app/confirm-order/upsert-receipt/page.tsx
@@ -6,8 +6,7 @@ import { useAuthContext } from '@/context/AuthContext'
 import { useCity } from '@/hooks/useCity'
 import { getAddress, upsertAddress } from '@/services/address'
 import { AddressData } from '@/types/common'
-import { useSearchParams } from 'next/navigation'
-import { useRouter } from 'next/navigation'
+import { useRouter, useSearchParams } from 'next/navigation'
 import React, { useEffect } from 'react'
 import { useImmer } from 'use-immer'
 import { format } from 'date-fns'
